AnimationLayerBuilderを分離
