+++ Packages/com.hhotatea.avatar-pose-library/Runtime/Model/ConstVariables.cs
@@ -23,5 +23,6 @@ namespace com.hhotatea.avatar_pose_library.model
         
         // 1つのIntパラメーターで管理するAnimationの最大数。
         public const int MaxAnimationState = 255;
+        public const int BoolFlagCount = 31;
     }
 }
